+++ bin/fn.ps1
@@ -1,4 +1,5 @@
 $fn = $args[0]
 
 emacs --batch --directory "$env:MODULE_DIR" `
+  --module-assertions `
   -l "$env:PROJECT_ROOT\test-module\tests\main.el" -f "$fn"