+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 - Added `GlobalRef`, which allows keeping Lisp values around without an `Env`.
 - Reduced indirection when calling common built-in subroutines through `Env`.
+- Replaced `lazy_static` dependency with `once_cell`.
 
 ## [0.12.3] - 2020-02-18
 - Added `Value::car`, `Value::cdr`.